requirements.txt
dicord.py
python-decouple